@@ -1,19 +1,37 @@
-source :gemcutter
-gem "rails", "2.3.5"
-gem "rack", "1.0.1"
-gem "rake", "0.8.7"
-gem "gem_plugin", "0.2.3"
-gem "mime-types", "1.15"
-gem "fastthread", "1.0.7"
-gem "mongrel", ">= 1.1.5"
-gem "mongrel_cluster", "1.0.5"
-gem "nokogiri", "1.4.2"
-gem "httpauth", "0.1"
-gem "jviney-acts_as_taggable_on_steroids", "1.1"
-gem "mysql"
-gem "libxml-ruby", "1.1.2"
+source :rubygems
+gem "bundler"
+
+gem "rails", "3.0.3"
+gem "rake"
+gem "gem_plugin"
+gem "mime-types"
+gem "fastthread"
+gem "nokogiri"
+gem "httpauth"
+gem "acts_as_taggable_on_steroids"
+gem "ruby-openid"
+gem "ruby-openid-apps-discovery"
+gem "rspec-rails", "2.2.0"
+gem "delayed_job"
+gem "dynamic_form"
+
+group :postgres do
+  gem "pg"
+end
+
+group :mysql do
+  gem "mysql"   ### assume in system gems
+end
+
+group :development do
+  gem "heroku"
+  gem "sqlite3-ruby", "1.3.1"
+  gem 'ruby-debug-base19' if RUBY_VERSION.include? "1.9"
+  gem 'ruby-debug-base' if RUBY_VERSION.include? "1.8"
+  gem "ruby-debug-ide"
+end
 
 group :test do
-  gem "rspec-rails", ">= 1.1.11"
-  gem "jasmine", "0.10.3.1"
+  gem "jasmine", "1.0.1.1"
+  gem "headless", "0.1.0"
 end
