@@ -1,4 +1,4 @@
-var defaultRefreshTimeout = 30;
+var defaultRefreshTimeout = 60;
 var currentTimeout = null;
 
 function scheduleRefresh() {
@@ -11,7 +11,7 @@ function scheduleRefresh() {
 }
 
 function refresh() {
-  var cimonitorUrl = "/cimonitor";
+  var cimonitorUrl = "/dashboard";
   if (document.location) {
     cimonitorUrl += document.location.search;// copy all the ?params and add them to the cimonitor URL
     window.frames[0].location = cimonitorUrl;
