@@ -1,11 +1,11 @@
-require File.dirname(__FILE__) + '/../spec_helper'
+require 'spec_helper'
 
 describe ProjectsController do
   describe "with no logged in user" do
     describe "all actions" do
       it "should redirect to the login page" do
         get :index
-        response.should redirect_to(new_login_path)
+        response.should redirect_to(login_path)
       end
     end
   end
@@ -19,6 +19,7 @@ describe ProjectsController do
       get :index
       response.should be_success
       assigns(:projects).should_not be_nil
+      assigns(:aggregate_projects).should_not be_nil
     end
 
     it "should respond to new" do
