@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../spec_helper'
+require 'spec_helper'
 
 describe Project do
   before(:each) do
@@ -24,6 +24,14 @@ describe Project do
         @project.should_not be_valid
         @project.errors[:feed_url].should_not be_nil
       end
+
+      it "should allow both http and https" do
+        @project.feed_url = "http://bar.com/foo.rss"
+        @project.should be_valid
+
+        @project.feed_url = "https://bar.com/foo.rss"
+        @project.should be_valid
+      end
     end
   end
   
@@ -42,4 +50,93 @@ describe Project do
       @project.project_name.should == "BAZ"
     end
   end
+
+  describe "status_parser" do
+    describe "with reported success" do
+      before(:each) do
+        @status_parser = @project.parse_project_status(CCRssExample.new("success.rss").read)
+      end
+
+      it "should return the link to the checkin" do
+        @status_parser.url.should == CCRssExample.new("success.rss").xpath_content("/rss/channel/item/link")
+      end
+
+      it "should return the published date of the checkin" do
+        @status_parser.published_at.should ==
+            Time.parse(CCRssExample.new("success.rss").xpath_content("/rss/channel/item/pubDate")) 
+      end
+
+      it "should report success" do
+        @status_parser.should be_success
+      end
+    end
+
+    describe "with reported failure" do
+      before(:each) do
+        @status_parser = @project.parse_project_status(CCRssExample.new("failure.rss").read)
+      end
+
+      it "should return the link to the checkin" do
+        @status_parser.url.should == CCRssExample.new("failure.rss").xpath_content("/rss/channel/item/link")
+      end
+
+      it "should return the published date of the checkin" do
+        @status_parser.published_at.should == Time.parse(CCRssExample.new("failure.rss").xpath_content("/rss/channel/item/pubDate"))
+      end
+
+      it "should report failure" do
+        @status_parser.should_not be_success
+      end
+    end
+
+  #  describe "with invalid xml" do
+  #    before(:all) do
+  #      @parser = XML::Parser.string(@response_xml =  "<foo><bar>baz</bar></foo>")
+  #      @response_doc = @parser.parse
+  #      @status_parser = StatusParser.status( "<foo><bar>baz</bar></foo>")
+  #    end
+  #
+  #    it "should report failure" do
+  #      @status_parser.should_not be_success
+  #    end
+  #  end
+  end
+
+
+  describe "building_parser" do
+    before(:each) do
+      @project = CruiseControlProject.new(:name => "Socialitis", :feed_url => "http://foo.bar.com:3434/projects/Socialitis.rss")
+    end
+    
+    context "with a valid response that the project is building" do
+      before(:each) do
+        @status_parser = @project.parse_building_status(BuildingStatusExample.new("socialitis_building.xml").read)
+      end
+
+      it "should set the building flag on the project to true" do
+        @status_parser.should be_building
+      end
+    end
+
+    context "with a valid response that the project is not building" do
+      before(:each) do
+        @status_parser = @project.parse_building_status(BuildingStatusExample.new("socialitis_not_building.xml").read)
+      end
+
+      it "should set the building flag on the project to false" do
+        @status_parser.should_not be_building
+      end
+    end
+
+    context "with an invalid response" do
+      before(:each) do
+        @status_parser = @project.parse_building_status("<foo><bar>baz</bar></foo>")
+      end
+
+      it "should set the building flag on the project to false" do
+        @status_parser.should_not be_building
+      end
+    end
+  end
+  
 end
